extracted first ERPs from new portable EEG system
r, based on his scalp plot I must have the wrong
montage/electrode-location file because the typical ERPs I expected are
to see are there, but they are at the wrong locations!  Either that or
Megan plugged them into the wrong places (but I doubt that is the
case!). So I have just emailed BrainProducts to ask them to provide me
the relevant montage file. Hopefully as soon as I have the correct
montage file the ERPs will look exactly as expected!